backend/requirements.txt
asgiref==3.3.4
crispy-bootstrap5==0.6
Django==3.2
django-appconf==1.0.5
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-imagekit==4.1.0
django-js-asset==1.2.2
django-rest-framework==0.1.0
django-uuslug==1.2.0
djangorestframework==3.12.4
pilkit==2.0
Pillow==8.2.0
python-slugify==5.0.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
